+++ php7.4-litespeed/52-cs-wp.sh
@@ -1,17 +1,6 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-echo >&2 "Setting CS WP Env..."
-
-if [ -z ${CS_AUTH_KEY} ]; then
-  echo >&2 "CS_AUTH_KEY not set, generating random key..."
-  export CS_AUTH_KEY=$(xxd -l24 -ps /dev/urandom | xxd -r -ps | base64 | tr -d = | tr + - | tr / _)
-fi
-
-# Ensure we always have the correct auth key.
-sed -i "s/CS_AUTH_KEY=.*/CS_AUTH_KEY=$CS_AUTH_KEY/g" /usr/local/lsws/conf/httpd_config.conf
-
-# if it does not exist, create it.
-grep -q -e 'CS_AUTH_KEY' /usr/local/lsws/conf/httpd_config.conf || sed -i "/extprocessor lsphp/a  env                     CS_AUTH_KEY=$CS_AUTH_KEY" /usr/local/lsws/conf/httpd_config.conf
+set -e
 
 if [ -f "/opt/cs-wordpress-plugin-main/cstacks-config.php" ]; then
   echo >&2 "Updating ComputeStacks integration with latest version..."
@@ -48,7 +37,6 @@ else
 */30 * * * * www-data /usr/bin/curl http://localhost/wp-cron.php?doing_wp_cron
 EOF
 fi
-
 
 echo >&2 "Updating litespeed configuration..."