+++ php8.1-litespeed/51-wordpress.sh
@@ -1,4 +1,6 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+set -e
 
 mkdir -p /var/www/html/wordpress && chown www-data:www-data /var/www/html/wordpress
 mkdir -p /var/www/logs && chown www-data:www-data /var/www/logs