+++ php8.2-nginx/startup/55-wordpress.sh
@@ -1,16 +1,19 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+set -e
 
 mkdir -p /var/www/html/wordpress && chown www-data:www-data /var/www/html/wordpress
 cd /var/www/html/wordpress
 
-# check if /etc/nginx/sites-available/multisite-subdirectory.conf exists, if not copy it to /var/www/nginx/multisite-subdirectory.conf.example
-if [ -f /etc/nginx/sites-available/multisite-subdirectory.conf ]; then
-  cp /etc/nginx/sites-available/multisite-subdirectory.conf /var/www/nginx/multisite-subdirectory.conf.example
+# re-seed volume with up to date examples
+if [ -f /opt/nginx/multisite-subdirectory.conf ]; then
+  cp /opt/nginx/multisite-subdirectory.conf /var/www/nginx/multisite-subdirectory.conf.example
 fi
-# check if /etc/nginx/sites-available/multisite-subdomain.conf exists, if not copy it to /var/www/nginx/multisite-subdomain.conf.example
-if [ -f /etc/nginx/sites-available/multisite-subdomain.conf ]; then
-  cp /etc/nginx/sites-available/multisite-subdomain.conf /var/www/nginx/multisite-subdomain.conf.example
+if [ -f /opt/nginx/multisite-subdomain.conf ]; then
+  cp /opt/nginx/multisite-subdomain.conf /var/www/nginx/multisite-subdomain.conf.example
 fi
+
+chown -R www-data:www-data /var/www/nginx
 
 wait_for_db() {
   counter=0
@@ -31,7 +34,6 @@ setup_db() {
   mysql -h $WORDPRESS_DB_HOST -u $WORDPRESS_DB_USER -p$WORDPRESS_DB_PASSWORD --skip-column-names -e "CREATE DATABASE IF NOT EXISTS $WORDPRESS_DB_NAME;"
 }
 
-# Install plugin function installs first plugin specified, e.g. 'install_plugin litespeed'
 install_plugin() {
   echo >&2 "Installing $1..."
   sudo -u www-data wp plugin install $1 --activate && echo "Plugin $1 installed and activated succesfully!" || "Installing $1 failed; check if plugin name is correctly specified"
@@ -51,14 +53,23 @@ if ! [ "$(ls -A)" ]; then
 if (isset($_SERVER['HTTP_X_FORWARDED_PROTO']) && $_SERVER['HTTP_X_FORWARDED_PROTO'] === 'https') {
   $_SERVER['HTTPS'] = 'on';
 }
-define( 'FS_METHOD', 'direct' ); 
-define( 'WP_MEMORY_LIMIT', '96M' );
-define( 'CS_PLUGIN_DIR', '/opt/cs-wordpress-plugin-main' );
-define( 'WP_AUTO_UPDATE_CORE', 'minor' );
 PHP
 
   echo >&2 "Installing WordPress..."
   sudo -u www-data wp core install --url=$WORDPRESS_URL --title="$WORDPRESS_TITLE" --admin_user=$WORDPRESS_USER --admin_password=$WORDPRESS_PASSWORD --admin_email="$WORDPRESS_EMAIL" --skip-email
+
+  sudo -u www-data wp config set WP_AUTO_UPDATE_CORE 'minor'
+  sudo -u www-data wp config set FS_METHOD 'direct'
+  sudo -u www-data wp config set WP_MEMORY_LIMIT '96M'
+
+  echo >&2 "Uninstall default plugins"
+  sudo -u www-data wp plugin is-installed akismet && sudo -u www-data wp plugin uninstall akismet
+  sudo -u www-data wp plugin is-installed hello && sudo -u www-data wp plugin uninstall hello
+
+  echo >&2 "Configuring nginx helper"
+  sudo -u www-data wp config set RT_WP_NGINX_HELPER_CACHE_PATH '/var/run/nginx-cache'
+  sudo -u www-data wp option add rt_wp_nginx_helper_options --format=json '{"enable_purge":"1","cache_method":"enable_fastcgi","unlink_files":"unlink_files","enable_map":null,"enable_log":null,"log_level":"INFO","log_filesize":"5","enable_stamp":null,"purge_homepage_on_edit":"1","purge_homepage_on_del":"1","purge_archive_on_edit":null,"purge_archive_on_del":null,"purge_archive_on_new_comment":null,"purge_archive_on_deleted_comment":null,"purge_page_on_mod":"1","purge_page_on_new_comment":"1","purge_page_on_deleted_comment":"1","redis_hostname":"127.0.0.1","redis_port":"6379","redis_prefix":"nginx-cache:","purge_url":"","redis_enabled_by_constant":0,"smart_http_expire_form_nonce":"2923f3260e"}' --skip-plugins --skip-themes
+  sudo -u www-data wp plugin install nginx-helper --activate --skip-plugins --skip-themes
 
   if [[ -z "$WORDPRESS_PLUGINLIST" ]]; then
     echo "Pluginlist-variable empty, skipping plugin installation."
@@ -86,14 +97,12 @@ PHP
     sudo -u www-data wp language core install $WORDPRESS_LANGUAGE
     sudo -u www-data wp language core activate $WORDPRESS_LANGUAGE
   fi
-  
-  echo "Installing CS plugin"
-  sudo -u www-data mkdir -p wp-content/mu-plugins
-  sudo -u www-data cp /opt/cs-wordpress-plugin-main/cstacks-config.php wp-content/mu-plugins/
 
 else
 
   echo "Setting DB Host"
   sudo -u www-data wp config set DB_HOST $WORDPRESS_DB_HOST
+  echo >&2 "Configuring nginx cache path"
+  sudo -u www-data wp config set RT_WP_NGINX_HELPER_CACHE_PATH '/var/run/nginx-cache'  
 
 fi