+++ php8.2-nginx/startup/56-cs-wp.sh
@@ -1,20 +1,15 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
-echo >&2 "Setting CS WP Env..."
+set -e
 
-if [ -z ${CS_AUTH_KEY} ]; then
-  echo >&2 "CS_AUTH_KEY not set, generating random key..."
-  export CS_AUTH_KEY=$(xxd -l24 -ps /dev/urandom | xxd -r -ps | base64 | tr -d = | tr + - | tr / _)
-fi
-
-# Ensure we always have the correct auth key.
-sed -i "s/env\[CS_AUTH_KEY\] = .*/env\[CS_AUTH_KEY\] = '$CS_AUTH_KEY'/g" /etc/php/7.4/fpm/pool.d/www.conf
+echo >&2 "Configuring ComputeStacks path"
+sudo -u www-data wp config set CS_PLUGIN_DIR '/opt/cs-wordpress-plugin-main'
 
 if [ -f "/opt/cs-wordpress-plugin-main/cstacks-config.php" ]; then
   echo >&2 "Updating ComputeStacks integration with latest version..."
   sudo -u www-data mkdir -p /var/www/html/wordpress/wp-content/mu-plugins
+  rm /var/www/html/wordpress/wp-content/mu-plugins/cstacks-config.php
   sudo -u www-data cp /opt/cs-wordpress-plugin-main/cstacks-config.php /var/www/html/wordpress/wp-content/mu-plugins/
-  sudo -u www-data wp --path=/var/www/html/wordpress config set CS_PLUGIN_DIR /opt/cs-wordpress-plugin-main
 fi
 
 echo >&2 "Configuring wordpress cron jobs..."